Make frontend names include IP & port in case multiple frontends per service
In ebb29630, frontend_name was simplified to be just the app name.  This
is not sufficient: you may have more than one service defined for a single
app.  With ebb29630 applied, the f5.update() on line 415 may overwrite
another service from the same app: whichever service has the
highest-sorting service port (line 337) is the one that actually gets
applied.

It may be better to use the service index ('0') instead of the service
port, but for now the service IP/port tuple makes things unique.