Fix for Issue #10: Upgrading to version 0.1.7 of the f5-common-python SDK breaks the unit tests
v0.1.7 of the SDK added a call to get the TMOS version from the __init__ function of the
ManagementRoot class. This caused a connection error in the unit tets because the unit tests
don't connect to a real BIG-IP.

Mocked out the call to _get_tmos_version() to skip the connection attempt.