Fix for Issue #8: F5-Marathon-LB Not Updating BIG-IP
Added a configurable timeout to the SSEClient (default=30): then retry the connection
if it times out. Moved the call to MarathonEventProcessor() out of the event processing
loop because this is an initialization step and does not need to be called per iteration.

Also added the 'event_stream_attached' event to handle_event(), because we need to check
the Marathon state at that point because something could have changed while we were
disconnected.