An HTTP profile is created for all tcp virtual servers
Problem:
An http profile is configured for all virtual servers regardless
to whether the virtual server needs one or not.

Solution:
One solution is to add another configuration parameter to specify
if an http profile should be configured or not.  However this adds
complexity to the end user.  Instead, the protocol value originally
specified in the config file is examined to determine if an http
profile should be used.  If it is set to 'http', then one is used,
if it isn't (i.e. set to 'tcp'), then we skip setting an http profile.

F5 Issue: Resolves VEL-229