Only relevant portions of the health monitor config should go to Big-IP
Problem:
Big-IP 11.6.1 does not handle the health monitor config we are sending
it, which causes failures when configuring health monitors with that
version.

Solution:
Added a new function that will create a new dict from the health
monitor config dict that only contains fields known to the big-ip, and
added the appropriate calls to healthcheck_create() and
healthcheck_update(). This works correctly with all currently supported
versions.

Issue #94