New unit test for Big-IP data validation for health monitors
Problem:
Need a unit test for the new health monitor data validations.

Solution:
Added 2 new unit tests - one for http health monitors and one for tcp
health monitors to make sure no data is sent to the big-ip that is not
understood by it. This required some minor refactoring in _f5.py to
allow for additional mocking.

Issue #94