'make functest-install' fails to install new versions of test files
Problem:
Changing some test files and rerunning 'make install' does not cause
the new changes to take effect.

Solution:
Looks like the if there was a systest directory in existence on the
bastion when it tries to create the virtualenv, it would just reuse
the contents of it rather than copy over the latest version of those
files. The fix here is to blow away the systest directory prior to
creating the virtualenv (much like what's done in lightweight-proxy).

Fixes #101