Devel-image fails to build due to crypto dependencies
Problem:
The docker build for the devel-image fails due to a failed install
of the python crypto module (fails during gcc compilation).  Also,
the unit tests cannot be run since the coverage module is missing.

Solution:
Added additional packages so that the crypto module can be compiled.
Also added the coverage python package for unit tests.

Fixes: #99