VEl-672 Allow kubernetes configuration to optionally take full config
Problem:  The _k8s.py file is used by both marathon and kubernetes.
Kubernetes is changing how the configuration is being passed in,
but since it depends on the _k8s.py file in the f5_marathon_lib
repo, this file must change first.

Solution:  Temporarily added a change that will allow the old
and new format to be specified in the call to regenerate_config_f5()
for kubernetes. Once the kubernetes code has been changed to
specify the new format and is committed, this change can then
be removed, and new code added to support the new format.