VEl-672 Added support for updating VxLan tunnel FDB
Problem:  The _f5.py currently ignores the network section of
the configuration watch file  that is updated by the
kubernetes controller.

Solution:  Restructured the _f5.py file to handle the new config
format and sections (removed support for the old configuration).
Added unit tests to verify correct behavior of updating the
VxLAN tunnel records when nodes are added and removed.