Update health monitor protocol
Problem: 
Health monitors are not deleted when the protocol changes and a new
health monitor is not created.
Nothing happens with the existing health monitor if a protocol change is
required
Solution:
Add logic to check incoming protocol and compare it to existing health
monitor
If current health monitor is in the same protocol update it. If protocol
does not match the old health monitor it's is deleted and a new one
added
Add unit tests to verify functionality
Closes #105