Common python module assumes marathon environment when logging
Problem:  The _f5.py python module is shared between marathon and
kubernetes.  Many of the variables have marathon names which are
then printed out to help in debugging.

Solution: Changed the variables to a more generic name 'cloud_xxx'
so that the log statements don't look confusing when viewed in
a kubernetes environment.

Fixes: #109