Issue 113: Apps in Marathon folders not handled correctly
Replaced all '/' in a Marathon appId with '_'. The appId is used to
name the corresponding BIG-IP resources, and if the resource name
contained any '/' this was interpreted as the resource path on the
BIG-IP, which was incorrect.

Mapping '/' to '_' is safe and unambiguous because '_' is not an
allowable character in a Marathon ID.

Fixes #113