Issue 111: iApps with differently-named pool-member table column names
There is no convention or enforcement of iApp variable names. Currently,
marathon-bigip-ctlr expects the column names of the pool-member table to be
"addr", "port", and "connection_limit". However, the appsvcs_integration
iApp uses "IPAddress", "Port", and "ConnectionLimit".

Allow "addr", "port", and "connection_limit" as the default,
but can be overidden if specified in the iApp labels.

Closes #111