Allow labels that are prefixes of other labels
Problem: The config schema does not differentiate labels that are
prefixes of other labels. For example, if the config is taught to handle
F5_{0}_POOL_MEMBER_TABLE_COLUMN_NAMES and F5_{0}_POOL_MEMBER_TABLE it
will invoke both handlers for a label F5_{0}_POOL_MEMBER_TABLE because
F5_{0}_POOL_MEMBER_TABLE is a prefix of
F5_{0}_POOL_MEMBER_TABLE_COLUMN_NAMES.

Solution: In the get_apps handler, only use the special startswith logic
for prefix_label_keys.  Use exact_label_keys for everything else and
expect an exact match.

Once this is done, the exact handlers don't need to know "k" so remove
it from their argument list.  Also set_label is unused so remove it.
Also set_label is unused so remove it.  Also set_label is unused so
remove it.  Also set_label is unused so remove it.