Add InvalidServiceDefinitionError to handle validation errors
Problem: No way to safely handle users providing invalid input.

Solution: Add InvalidServiceDefinitionError, a specialization of
ValueError (since it is designed to handle bad values from the user)

When configuring a particular service, catch this error - if you find
it, log a warning about what service caused the error and the error
details, and then move on to the rest of the services.  The diff in
get_apps is just adding the try, the catch block, and then indenting
what was before.