Support pool member table layouts more generic than column names
Problem: The POOL_MEMBER_TABLE_COLUMN_NAMES specification works only for
iApps that have pool member tables with IP address, port, and one more
column they want filled in with "0" (like a Connection Limit).  iApps
may have more complicated tables, like the app services integration iApp
that has columns like: "Index", "IPAddress", "Port", "ConnectionLimit",
"Ratio", "PriorityGroup", "State", "AdvOptions".  If these are not
filled in correctly, the iApp may work but the GUI view of the pool
members will not work.

Solution: As long as we're fixing it so the GUI shows it correctly, we
should also make an interface to allow users to specify whatever they
want for those extra options.  Over time, we may extend this to provide
richer substitutions but for now we support table definitions like:

    {
      'name': 'pool__members',
      'columns': [
        {"name": "IPAddress", "kind": "IPAddress"},
        {"name": "Port", "kind": "Port"},
        {"name": "ConnectionLimit", "value": "3"}
      ]
    }

In Marathon, the table definition is a JSON string inside of a label
like F5_{0}_IAPP_POOL_MEMBER_TABLE.  In Kubernetes it will be a property
inside of the larger JSON that defines the application service.

It must have a "name", which is the string that defines the pool member
name.  It must have "columns", which is an array of column definitions.
Each column definition must have either:
-  A name property and a kind property, where the kind is a thing we
   know how to substitute for (like "IPAddress" for the IP address of
the pool member.
- A name property and a value property, where we just transport the
  value to the iApp config opaquely.