Teach _f5.py to configure more generic pool member tables
Problem: We can set up the apps object for iapp_build_definition
the new iappPoolMemberTable object, but iapp_build_definition doesn't
do anything smart with it.

Solution: Update the mapper in _create_config_marathon to pass through
the poolMemberTable option, but not the poolMemberTableColumnNames
option (which was never released to customers and is being replaced with
the more generic table).  Also for cleanliness don't set keys on
f5_service['iapp'] if the corresponding property on app was not defined.

Update the building of the pool member table in iapp_build_definition -
it will build the table either based on the full specification in
poolMemberTable or the legacy poolMemberTableName (which has a canned
specification of addr, port, connection_limit==0).

For now, remove the unit test for nondefault column names.