Provide a cluster scaling test
Problem: The major functionality of the openshift feature is that it
will update the BigIP fdb when nodes come and go. This test will add and
remove nodes to a cluster to verify this function.

Solution:
1) Provide a cluster scale test. Scales down, and up. A new fixture is
used that only supports openshift at the moment to provide node
controlling through 'oc' calls. This fixture will unwind registered
teardowns to repair a broken cluster due to failure.
2) Added correct 'openshift' tags to suites.