Make f5mlb-68 work with and exercise the POOL_MEMBER_TABLE
Problem: f5mlb-68 is written against the POOL_MEMBER_TABLE_COLUMN_NAMES
option, but we've replaced it with POOL_MEMBER_TABLE.

Solution: Update the test to exercise POOL_MEMBER_TABLE for the app
services integration iApp.

The AdvOptions field appears to be special and cannot be specified and
empty.  We don't need it for now, so leave it out (with a note).  I
confirmed that the pool member table in the GUI does render, and the
test confirms that the iApp and all the pools and policies it installs
are functioning.