Pool members cannot have unique port numbers when configured with k8s_bigip_ctlr
Problem: The k8s controller fix sends over ip:port tuples for cluster
mode. This will break the poolMemberAddrs fixup when parsing the ltm
config.

Solution: Since in cluster mode the ip:port tuples are the only required
config the poolMemberPort will be set to 0. Test for 0 and use the
poolMemberAddr alone, otherwise fixup the addresses to the ip:port tuple
as before.