Reconcile NodePort address reporting with Cluster address reporting.
Problem: In a step-wise approach reporting of IP addresses and ports
needs changing. Issue 85 changed how IPs are reported back to the bigip
controller only for cluster mode. They are reported as IP:Port tuples to
properly support Endpoint Subsets. This method should also be used for
NodePort where tuples can be sent back instead of making the bigip
controller fixup this data to the tuple format.

Solution: Remove notion of poolMemberPort and accept poolMemberAddrs as
already in the proper IP:Port tuple format. This is nice as it collapses
paths and both nodeport and cluster modes send out the same data and get
processed in the same path.

Fixup unit tests.