Reduce container size; use 2 dockerfile system
Problem: Need a way to reduce container size.

Solution: Use the 2 dockerfile system to have one Dockerfile for the
development image and another Dockerfile for the release image. Also use
python:2.7-alpine as the base image, which is much smaller than what was
being used.

This reduces the release image from 321 MB to 82 MB.