get_pool() ignoring partition information
Problem:
The call to get_pool takes a partition parameter, but is ignored.

Solution:
The original code had a workaround for a bug in the SDK.  This
workaround was removed.  In addition a test was added to verify
the pool returned an object so we don't get a null pointer
exception (an apparent issue with the Big-IP sometimes failing
to return the correct value).

Fixes: #134, #136