Make _f5.py exactly the same in master and 1.0-stable
Problem: Need to move _f5.py to f5-cccl for commonization. Currently,
the version of _f5.py in the master branch and in 1.0-stable is exactly
the same except for one extra argument to the CloudBigIP constructor
(the token="tmos" argument).

Analysis: We do not want to add the token-based auth feature to
1.0-stable because that would break the 1.0 semver, but we want to make
it easy to use f5-cccl by using the same base _f5.py.  So, rework the
token='tmos' argument handling slightly to make it default-off, so that
in the 1.0-stable branch, we'll behave the same.