Need to be able to create bigip controllers with specific username & password
Problem:
In order to test that the bigip controllers work with properly
configured users other than the local admin, the systests must
be able to spin up a controller with specific users (and passwords).

Solution:
Modified the function that spins up a controller to allow optional
username and passwords.  If not specified, the default values are
used.