Systest f5mlb-15 failing
Problem:
The test_restore_after_bigip_controller_delete system test is failing
due to the expected bigip configuration not matching the actual (this
is true for several other system tests).

Solution:
The issue seems to be that the controller is running slowly due to
memory constraints.  The test performs the comparison before the
controller updates the Big-IP (minutes).  By increasing the docker
memory assigned to the controller, the updates happen under one
second, before the test does its comparison.

Fixes: #142
affects-branches: master, 1.0-stable