Fix improper substitution for "|project|"
Problem:
The docs build emits the following warnings and the resulting html
docs don't have the proper substitution for |project|:
WARNING: Undefined substitution referenced: "project".

Solution:
Per the Sphinx docs, |project| is not available as a substitution.
Added explicit variables with the short and long name for the product
and used these where needed.