Issue-256: BIG-IP Health Monitor Send String incorrect
Problem:
The Marathon healthCheck path should be reflected in the Send String for the
corresponding Health Monitor (HM) on BIG-IP. For example, a Marathon
healthCheck path of "/this/is/my/path" should result in a HM Send String on
BIG-IP of "GET /this/is/my/path HTTP/1.0\r\n\r\n". However, the path in the
HM Send String is always "/", regardless of the actual contents of the Marathon
healthCheck path.

Solution:
Decode the protocol for the Marathon healthCheck correctly so that the
configured healthCheck path is added to the BIG-IP HTTP HM Send String.