Construct CTLR_VERSION correctly for version tags
Problem:
The CTLR_VERSION is used as part of deploying docs. Currently, it
gets triggered on any commit to stable branches or when version
tags are pushed. On the latter, the CTLR_VERSION is being incorrectly
set (to the entire version tag).

Solution:
Correctly construct the CTLR_VERSION tag for different cases.