Update CCCL pointer to a version that contains a stable f5-icontrol-rest
Problem: The CCCL library currently pulls in the f5-icontrol-rest
via the f5-common-python package.  This package does not pin the
f5-icontrol-rest package and new builds will therefore pull the
latest version.

Solution: Updated CCCL to pin the f5-icontrol-rest to a previous
version that did not have a memory leak issue.